BACKENDS: renamed stacktraces to stacktrace and fixed includes
also remove the c++11 check from configure as we now compile with c++11 anyway
